teams_data/requirements.txt
requests
bs4
pandas
#selenium==3.141.0
#chromedriver-binary==84.0.4147.30
constants==0.6.0
google-cloud-storage
google-cloud-bigquery
lxml
flask
datetime